Read in intersect config from yaml
